refactor(tree): move IdAllocator to utils
