+++ experimental/dds/tree2/src/feature-libraries/sequence-field/rebase.ts
@@ -5,12 +5,11 @@
 
 import { assert, unreachableCase } from "@fluidframework/core-utils";
 import { StableId } from "@fluidframework/runtime-definitions";
-import { fail } from "../../util";
+import { IdAllocator, fail } from "../../util";
 import { ChangeAtomId, ChangesetLocalId, RevisionTag, TaggedChange } from "../../core";
 import {
 	CrossFieldManager,
 	CrossFieldTarget,
-	IdAllocator,
 	NodeExistenceState,
 	RevisionMetadataSource,
 } from "../modular-schema";
